new file:   Resources/Encodings/mcp_ogg_112kbit_vorbis_96kbit_32000.plist 	renamed:    lib/mediacastproducer/actions.rb -> lib/mcp/actions.rb 	renamed:    lib/mediacastproducer/actions/faststart.rb -> lib/mcp/actions/faststart.rb 	renamed:    lib/mediacastproducer/actions/reference.rb -> lib/mcp/actions/reference.rb 	renamed:    lib/mediacastproducer/actions/transcode.rb -> lib/mcp/actions/transcode.rb 	renamed:    lib/mediacastproducer/actions/videosize.rb -> lib/mcp/actions/videosize.rb 	renamed:    lib/mediacastproducer/constants.rb -> lib/mcp/constants.rb 	renamed:    lib/mediacastproducer/qt/qt.rb -> lib/mcp/qt/qt.rb

+++ lib/mcp/qt/qt.rb
@@ -12,7 +12,7 @@
 require 'fileutils'
 require 'common/pcast_exception'
 require 'qt/qt'
-require 'mediacastproducer/constants'
+require 'mcp/constants'
 
 MOVIE_DIMENSION_FLAVORS = { :track => OSX::QTTrackDimensionsAttribute, 
                             :clean => OSX::QTMovieApertureModeClean,
@@ -104,9 +104,9 @@ class McastQT < PcastQT
           log_error("Failed to restore chapters") 
         end
       end
-      FileUtils.mv(non_chapterized_output_path, output_path)
     end
 
+    FileUtils.mv(non_chapterized_output_path, output_path) if File.exist?(non_chapterized_output_path)
     return true
   end
